requirements.txt
discord.py
requests
selenium
webdriver-manager